Checkstyle status (#7)
* Added three additional checkstyle status checks

Added three check stylestatus check rules to the checkstyle_checks.xml file.  The first status check added is LineLength which simply checks for long lines that can make things more difficult to read, it is currently set to a max of 100 characters long but this can be changed if that is too small I just used the value that was in the checkstyle file provided to us.  The second status check added is GenericWhitespace which checks to make sure that the whitespace around Generic tokens (angle brackets) "<" and ">" are formatted correctly and follow the typical convention.  The third status check added is Indentation to check that the indentation of the java code we write is the correct format and standardized between whoever is writing code.

* Remove GenericWhitespace module from checkstyle

Removed one of the status checks so that it passes the test for checkpoint one because there can only be two added.

* Restore LineLength and Indentation modules in config

* Fix XML structure by removing extra closing tag

* Remove duplicate LineLength and Indentation modules

* Update Checkstyle configuration for line length

* Refactor Indentation module into TreeWalker

Moved Indentation module configuration to TreeWalker module.